feat(btc): support btc on lite
