+++ packages/coin-btc/package.json
@@ -1,10 +1,11 @@
 {
   "name": "@coolwallet/btc",
-  "version": "1.1.9",
+  "version": "2.0.0-beta.0",
   "description": "Coolwallet Bitcoin sdk",
   "main": "lib/index.js",
   "scripts": {
-    "test": "NODE_ENV=development jest",
+    "test": "NODE_ENV=development jest --runInBand",
+    "test-lite": "NODE_ENV=development CARD=lite jest --runInBand",
     "ci-test": "NODE_ENV=production jest --runInBand",
     "build:types": "tsc --emitDeclarationOnly",
     "build:ts": "babel src --out-dir lib --extensions \".ts,.tsx\" --source-maps inline",
@@ -27,22 +28,19 @@
   ],
   "author": "coolwallet-team",
   "license": "ISC",
-  "peerDependencies": {
-    "@coolwallet/core": "^1.1.23"
-  },
   "devDependencies": {
     "@babel/cli": "^7.7.0",
     "@babel/core": "^7.7.0",
     "@babel/preset-env": "^7.7.1",
     "@babel/preset-typescript": "^7.7.7",
-    "@coolwallet/core": "^1.1.23",
     "ecpair": "^2.1.0"
   },
   "dependencies": {
     "@bitcoin-js/tiny-secp256k1-asmjs": "^2.2.3",
     "@types/bn.js": "^5.1.1",
     "bitcoinjs-lib": "^6.1.5",
-    "bn.js": "^5.2.1"
+    "bn.js": "^5.2.1",
+    "@coolwallet/core": "2.0.0-beta.15"
   },
   "repository": {
     "type": "git",
